+++ spec/models/event_spec.rb
@@ -5,6 +5,8 @@ RSpec.describe Event, type: :model do
   it { should belong_to :fraction }
   it { should have_many(:subscribes).dependent(:destroy) }
   it { should have_many(:characters).through(:subscribes) }
+  it { should have_many(:signed_subscribes).class_name('Subscribe') }
+  it { should have_many(:signed_characters).through(:signed_subscribes).source(:character) }
 
   it 'factory should be valid' do
     event = build :event